% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxt3.R
\docType{package}
\name{Maxt3}
\alias{Maxt3}
\title{Find test statistic value and corresponding critical value}
\description{
This package finds critical value of Max-T test for testing equality
of means against ordered alternatives in one-way ANOVA model
with heteroscedasticity when number of groups is 3.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
