% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxt3crit.R
\name{Maxt3crit}
\alias{Maxt3crit}
\title{Find test statistic value and corresponding critical value}
\usage{
Maxt3crit(g1, g2, g3, alpha)
}
\arguments{
\item{g1}{a real vector}

\item{g2}{a real vector}

\item{g3}{a real vector}

\item{alpha}{real number between 0 and 1}
}
\value{
Numerical vector
}
\description{
More detailed description
}
\examples{
g1=c(0,1,2,4,2,1,3,4)
g2=c(1,2,4,6,1,6,0,-2,-5)
g3=c(1,2,4,0,1,4,6,2,6)
Maxt3crit(g1,g2,g3,0.05)

}
